Re-arranged installation script.
When using the script, Caleb was having the issue that pip was
trying to install mpi4py and was unable to do so, therefore
being unable to finish the mcdc installation. Installing pip
with conda first made it so pip no longer tried to install mpi4py.
Possibly an issue exclusive to osx-arm64 machines.